Simplify Header further
